make vs code gpg key install idempotent

+++ setup/setup.d/30-vscode.sh
@@ -3,10 +3,12 @@
 notice "Installing VS Code Repo"
 
 # https://code.visualstudio.com/docs/setup/linux#_debian-and-ubuntu-based-distributions
-curl https://packages.microsoft.com/keys/microsoft.asc \
-  | gpg --dearmor \
-  > "$TMP/packages.microsoft.gpg"
-install -o root -g root -m 644 "$TMP/packages.microsoft.gpg" /etc/apt/trusted.gpg.d/
+if [[ ! -f /etc/apt/trusted.gpg.d/packages.microsoft.gpg ]]; then
+  curl https://packages.microsoft.com/keys/microsoft.asc \
+    | gpg --dearmor \
+    > "$TMP/packages.microsoft.gpg"
+  install -o root -g root -m 644 "$TMP/packages.microsoft.gpg" /etc/apt/trusted.gpg.d/
+fi
 
 echo "deb [arch=amd64 signed-by=/etc/apt/trusted.gpg.d/packages.microsoft.gpg] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list
